+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * return exit code 1 if at least one task failed (#12)
 * allow selecting slaves with constraints (#6)
 * allow to specify mesos role (#5)
+* find mesos master with zookeeper leader detector (#9)
 
 ## v0.1.0